Bug: 81234 - View Mail missing from open source upgrade
View Mail code was extracted out of the Admin Console code into an extension. It was created in ZimbraNetwork, as a result of which the FOSS users were no longer able to access it. Moving back the View Mail extension from ZimbraNetwork to FOSS ZimbraAdminExt and updating build scripts.

http://bugzilla.zimbra.com/show_bug.cgi?id=81234

Copied from Perforce
 Change: 459753